import alias changed to @/<folder name> inside src
