+++ src/components/button/DeleteNoteButton.tsx
@@ -1,6 +1,6 @@
 import { FaRegTrashCan } from "react-icons/fa6";
-import { ActionButton, ActionButtonProps } from ".";
-import { deleteNoteAtom } from "../../store";
+import { ActionButton, ActionButtonProps } from "@/components";
+import { deleteNoteAtom } from "@/store";
 import { useSetAtom } from "jotai";
 
 export const DeleteNoteButton = ({ ...props }: ActionButtonProps) => {
