+++ src/components/button/index.ts
@@ -1,3 +1,4 @@
 export * from "./ActionButton";
 export * from "./DeleteNoteButton";
 export * from "./NewNoteButton";
+export * from "./OpenFolderButton";
