+++ src/components/index.ts
@@ -3,4 +3,5 @@ export * from "./button";
 
 export * from "./TitleBar";
 export * from "./NotePreviewList";
+export * from "./NotePreview";
 export * from "./Editor";
