+++ src/models/config.ts
@@ -1,0 +1,3 @@
+export type Obsidian = {
+  lastOpenedDir: string;
+};
