+++ src/models/index.ts
@@ -1,11 +1,2 @@
-export type NoteInfo = {
-  title: string;
-  path: string;
-  lastEditTime: number;
-};
-
-export type NoteContent = string;
-
-export type Obsidian = {
-  lastOpenedDir: string;
-};
+export * from "./config";
+export * from "./note";
