+++ src/models/note.ts
@@ -1,0 +1,7 @@
+export type NoteInfo = {
+  title: string;
+  path: string;
+  lastEditTime: number;
+};
+
+export type NoteContent = string;
