+++ src/utils/index.ts
@@ -1,16 +1,2 @@
-import clsx, { ClassValue } from "clsx";
-import { twMerge } from "tailwind-merge";
-
-const dateFormatter = new Intl.DateTimeFormat(
-  Intl.DateTimeFormat().resolvedOptions().locale,
-  {
-    dateStyle: "short",
-    timeStyle: "short",
-  },
-);
-
-export const formatDateFromMs = (ms: number) => dateFormatter.format(ms);
-
-export const cn = (...args: ClassValue[]) => {
-  return twMerge(clsx(...args));
-};
+export * from "./fs";
+export * from "./utils";
