+++ src/utils/utils.ts
@@ -1,0 +1,16 @@
+import clsx, { ClassValue } from "clsx";
+import { twMerge } from "tailwind-merge";
+
+const dateFormatter = new Intl.DateTimeFormat(
+  Intl.DateTimeFormat().resolvedOptions().locale,
+  {
+    dateStyle: "short",
+    timeStyle: "short",
+  },
+);
+
+export const formatDateFromMs = (ms: number) => dateFormatter.format(ms);
+
+export const cn = (...args: ClassValue[]) => {
+  return twMerge(clsx(...args));
+};
