+++ tsconfig.json
@@ -18,7 +18,16 @@
     "strict": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
-    "noFallthroughCasesInSwitch": true
+    "noFallthroughCasesInSwitch": true,
+
+    "baseUrl": ".",
+    "paths": {
+      "@/components": ["./src/components"],
+      "@/hooks": ["./src/hooks"],
+      "@/models": ["./src/models"],
+      "@/store": ["./src/store"],
+      "@/utils": ["./src/utils"]
+    }
   },
   "include": ["src"],
   "references": [{ "path": "./tsconfig.node.json" }]
