+++ vite.config.ts
@@ -12,4 +12,13 @@ export default defineConfig(async () => ({
       ignored: ["**/src-tauri/**"],
     },
   },
+  resolve: {
+    alias: {
+      "@/components": resolve(__dirname, "./src/components"),
+      "@/hooks": resolve(__dirname, "./src/hooks"),
+      "@/models": resolve(__dirname, "./src/models"),
+      "@/store": resolve(__dirname, "./src/store"),
+      "@/utils": resolve(__dirname, "./src/utils"),
+    },
+  },
 }));
